Skills/Stock/HongKong_Stock/ma-pullback-strategy/requirements.txt
akshare>=1.12.0
pandas>=1.5.0
numpy>=1.23.0
yfinance>=0.2.0
pyyaml>=6.0
